Created a parent class for food enemies and implemented their moving
Since the food enemies have similar functions, I created a parent class for them with functions such as wandering, fleeing from the player, obstacle colission. What is left is to implement them being hurt and killed. For some of them, maybe we will make a special animation for their moving, but in the future